Added OhjeetUudelleOpiskelijalle to UI
Unit test for controller
To prove that ohjeetUudelleOpiskelijalle propagates
from backend to controller
Fixed related e2e tests
Remove not-needed directive parameter
Remove not used fixture
ohjeetUudelleOpiskelijalle field to applications api result

applications structure changed

Refactored "tarjonta path". Now the link to
instructions is not stored in hakutoiveet, but in
hakemus, in its own map "ohjeetUudelleOpiskelijalle".
This way it will be possible to support also the
"ataru path" in which hakutoiveet list is not sent
to the frontend. Yet, the frontend will be able to
get the link from "ohjeetUudelleOpiskelijalle" map.

Ataru tests and instructions to new students

Refactored the Ataru path to be testable.
Added ohjeetUudelleOpiskelijalle also to
Ataru path. Support also for non-secure scenarios

Showing the link to instrucitons in more cases

Better handle the case when hakukohdeOid is not found
in hakemusData

Open links in new window/tab

OhjeetUudelleOpiskelijalle, Oili, ja Kela links are covered

Fixes after rebase