OY-300 don't use scalatra sessions
It turned out that the database reads for SessionInfo are so
lightweight that there is no point in trying to optimize by keeping
SessionInfo in memory within scalatra session field "sessionInfo".

With this change, we are replacing the use of scalatra sessions with
direct read from database (SessionInfoRetriever). It means that for many
requests, there will be two session lookups in database, but it is ok,
because database will in most cases read the data from some sort of cache
memory, and will not require any additional UI operation.